move logger declaration
